+++ spec/ramda_spec.rb
@@ -9,6 +9,7 @@ describe Ramda do
 
   # r(:keys_in)
   # r(:math_mod)
+  # r(:to_pairs_in)
   # r(:values_in)
   r(:add)
   r(:all)